# cmd/docs/swagger.yaml
basePath: /
definitions:
  controllers.ErrorResponse:
    properties:
      details:
        description: Additional details about the error
        example: Invalid input data
        type: string
      error:
        description: Error message
        example: Bad Request
        type: string
    type: object
  controllers.InternalErrorResponse:
    properties:
      details:
        description: Additional details about the error
        example: Server Error
        type: string
      error:
        description: Error message
        example: Internal Server Error
        type: string
    type: object
  controllers.InvalidAPIKeyResponse:
    properties:
      details:
        description: Additional details about the error
        example: API key is invalid
        type: string
      error:
        description: Error message
        example: Invalid API key
        type: string
    type: object
...
  controllers.NotFoundResponse:
    properties:
      details:
        description: Additional details about the error
        example: No Data found for the input provided
        type: string
      error:
        description: Error message
        example: No data
        type: string
    type: object
...
  models.Organisation:
    description: Organisation model containing all organisation-related information.
    properties:
      org_id:
        description: Organisation ID
        example: "12345"
        type: string
      org_name:
        description: Organisation Name
        example: Acme Corp
        type: string
      org_uuid:
        description: Auto-generated UUID
        example: uuid-v4
        type: string
      status:
        allOf:
        - $ref: '#/definitions/models.OrganisationStatus'
        description: Organisation Status
        example: Active
    type: object
  models.OrganisationReq:
    description: OrganisationReq model containing all organisation request related
      information.
    properties:
      org_id:
        description: Organisation ID
        example: "12345"
        type: string
      org_name:
        description: Organisation Name
        example: Acme Corp
        type: string
      status:
        allOf:
        - $ref: '#/definitions/models.OrganisationStatus'
        description: Organisation Status
        example: Active
    type: object
  models.OrganisationStatus:
    enum:
    - Created
    - Active
    - InActive
    type: string
    x-enum-varnames:
    - OrgCreated
    - OrgActive
    - OrgInActive
...
host: localhost:9000
info:
  contact: {}
  description: This is the API documentation for the Fverify backend.
  title: FVerify API
  version: "1.0"
paths:
  /api/v1/organisations:
    get:
      consumes:
      - application/json
      description: Retrieve all organisations in the system
      parameters:
      - description: API key
        in: header
        name: X-API-Key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Organisation'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.InvalidAPIKeyResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.NotFoundResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.InternalErrorResponse'
      summary: Get all organisations
      tags:
      - Organisations
    post:
      consumes:
      - application/json
      description: Create a new organisation in the system
      parameters:
      - description: API key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: Organisation data
        in: body
        name: organisation
        required: true
        schema:
          $ref: '#/definitions/models.OrganisationReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Organisation'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.InvalidAPIKeyResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.InternalErrorResponse'
      summary: Create a new organisation
      tags:
      - Organisations
  /api/v1/organisations/{org_id}:
    put:
      consumes:
      - application/json
      description: Update an existing organisation's details
      parameters:
      - description: API key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: Organisation ID
        in: path
        name: org_id
        required: true
        type: string
      - description: Updated organisation data
        in: body
        name: organisation
        required: true
        schema:
          $ref: '#/definitions/models.OrganisationReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Organisation'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.InvalidAPIKeyResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.NotFoundResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.InternalErrorResponse'
      summary: Update an organisation
      tags:
      - Organisations
...
schemes:
- http
swagger: "2.0"